fix: ctx処理をhttp共通の関心毎として定義
